pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nami"
version = "0.1.0"
description = "Gesture-controlled audio modulation via MIDI"
requires-python = ">=3.11"
dependencies = [
    "mediapipe>=0.10",
    "opencv-python>=4.8",
    "mido[ports-rtmidi]>=1.3",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
nami = "nami.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/nami"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
